Fix bus-downloader compatibility with async/await
