+++ Sources/App/Models/Bus.swift
@@ -124,58 +124,6 @@ extension Collection where Element == Bus {
 	
 }
 
-extension Set where Element == Bus {
-	
-	/// Download the latest system bus data.
-	/// - Parameters:
-	///   - application: The current application object.
-	///   - busesCallback: A callback that's given a `Set<Bus>` instance with new bus objects. Note that these bus objects will **not** contain any user-reported location or congestion data and therefore must be separately merged with any existing bus data.
-	static func download(application: Application) async throws -> Set<Bus> {
-		let rawString = try String(contentsOf: Constants.datafeedURL)
-		let buses = rawString.split(separator: "\r\n")
-			.dropFirst()
-			.dropLast()
-			.compactMap { (rawLine) -> Bus? in
-				guard let backendIDRange = rawLine.range(of: #"(?<=(Vehicle\sID:))\d+"#, options: [.regularExpression]) else {
-					return nil
-				}
-				guard let latitudeRange = rawLine.range(of: #"(?<=(lat:))-?\d+\.\d+"#, options: [.regularExpression]), let latitude = Double(rawLine[latitudeRange]) else {
-					return nil
-				}
-				guard let longitudeRange = rawLine.range(of: #"(?<=(lon:))-?\d+\.\d+"#, options: [.regularExpression]), let longitude = Double(rawLine[longitudeRange]) else {
-					return nil
-				}
-				guard let timeRange = rawLine.range(of: #"(?<=(time:))\d+"#, options: [.regularExpression]), let dateRange = rawLine.range(of: #"(?<=(date:))\d{8}"#, options: [.regularExpression]) else {
-					return nil
-				}
-				let backendID = String(rawLine[backendIDRange])
-				let id = Buses.sharedInstance.busIDMap[backendID]
-				let formatter = DateFormatter()
-				formatter.dateFormat = "HHmmss'|'MMddyyyy"
-				formatter.timeZone = TimeZone(abbreviation: "UTC")!
-				let dateString = "\(rawLine[timeRange])|\(rawLine[dateRange])"
-				guard var date = formatter.date(from: dateString) else {
-					return nil
-				}
-				let coordinate = Coordinate(latitude: latitude, longitude: longitude)
-				if date > Date() {
-					let oldBus = try? await Bus.query(on: application.db)
-						.filter(\.$id == id)
-						.first()
-					if let oldLocation = oldBus?.locations.resolved, oldLocation.coordinate == coordinate {
-						date = oldLocation.date
-					} else {
-						date = Date.now
-					}
-				}
-				let location = Bus.Location(id: UUID(), date: date, coordinate: coordinate, type: .system)
-				return Bus(id: id, locations: [location])
-			}
-		return Set(buses)
-	}
-	
-}
-
 extension Collection where Element == Bus.Location {
 	
 	/// The resolved location datum from the bus's GPS hardware.
